+++ lib/hux.dart
@@ -49,6 +49,7 @@ export 'src/components/cards/hux_card.dart';
 export 'src/components/inputs/hux_input.dart';
 export 'src/components/inputs/hux_date_input.dart';
 export 'src/components/inputs/hux_checkbox.dart';
+export 'src/components/inputs/hux_radio.dart';
 export 'src/components/switch/hux_switch.dart';
 export 'src/components/badge/hux_badge.dart';
 export 'src/components/feedback/hux_alert.dart';